chore: move k8s tools for testing to their own crate
