+++ k8s-test-env/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "k8s-test-env"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+serde = { workspace = true, features = ["derive"] }
+serde_yaml = { workspace = true }
+serde_json = { workspace = true }
+schemars = { workspace = true }
+futures = { version = "0.3.28"}
+kube = { version = "0.90.0", features = ["runtime", "derive"]}
+k8s-openapi = { version = "0.21.1", features = ["latest"]}
+tokio = { version = "1.35.1", features = [
+  "rt-multi-thread",
+  "macros",
+] }
