+++ k8s-test-env/src/lib.rs
@@ -1,0 +1,8 @@
+//! k8s-test-env
+//!
+//! This crate provides tools to perform tests requiring a k8s cluster running.
+//!
+
+pub mod env;
+pub mod foo_crd;
+pub mod runtime;
