+++ super-agent/Cargo.toml
@@ -74,6 +74,7 @@ tokio = { version = "1.32.0", features = [
 ] }
 wiremock = "0.6.0"
 prost = "0.11.8"
+k8s-test-env ={ path = "../k8s-test-env" }
 
 [build-dependencies]
 glob = "0.3.1"
